+++ sdk/inc/azure/iot/az_iot_provisioning_client.h
@@ -218,6 +218,22 @@ typedef struct
 } az_iot_provisioning_client_registration_state;
 
 /**
+ * @brief Azure IoT Provisioning Service operation status.
+ *
+ */
+typedef enum
+{
+  // Device assignment in progress.
+  AZ_IOT_PROVISIONING_STATUS_UNASSIGNED,
+  AZ_IOT_PROVISIONING_STATUS_ASSIGNING,
+
+  // Device assignment operation complete.
+  AZ_IOT_PROVISIONING_STATUS_ASSIGNED,
+  AZ_IOT_PROVISIONING_STATUS_FAILED,
+  AZ_IOT_PROVISIONING_STATUS_DISABLED,
+} az_iot_provisioning_client_operation_status;
+
+/**
  * @brief Register or query operation response.
  *
  */
@@ -228,12 +244,8 @@ typedef struct
                          * (which may require several requests) is available only through
                          * #operation_status.  */
   az_span operation_id; /**< Operation ID of the register operation. */
-  az_span operation_status; /**< An #az_span containing the status of the register operation.
-                             * @details This can be one of the following: `unassigned`,
-                             * `assigning`, `assigned`, `failed`, `disabled`.
-                             * az_iot_provisioning_client_parse_operation_status() can optionally
-                             * be used to convert this into
-                             * the #az_iot_provisioning_client_operation_status enum. */
+  az_iot_provisioning_client_operation_status
+      operation_status; /**< An #az_iot_provisioning_client_operation_status of the register operation. */
   uint32_t retry_after_seconds; /**< Recommended timeout before sending the next MQTT publish. */
   az_iot_provisioning_client_registration_state
       registration_state; /**< If the operation is complete (success or error), the
@@ -259,33 +271,18 @@ AZ_NODISCARD az_result az_iot_provisioning_client_parse_received_topic_and_paylo
     az_iot_provisioning_client_register_response* out_response);
 
 /**
- * @brief Azure IoT Provisioning Service operation status.
- *
- */
-typedef enum
-{
-  // Device assignment in progress.
-  AZ_IOT_PROVISIONING_STATUS_UNASSIGNED,
-  AZ_IOT_PROVISIONING_STATUS_ASSIGNING,
-
-  // Device assignment operation complete.
-  AZ_IOT_PROVISIONING_STATUS_ASSIGNED,
-  AZ_IOT_PROVISIONING_STATUS_FAILED,
-  AZ_IOT_PROVISIONING_STATUS_DISABLED,
-} az_iot_provisioning_client_operation_status;
-
-/**
  * @brief Returns the #az_iot_provisioning_client_operation_status of a
  * #az_iot_provisioning_client_register_response object.
  *
- * @param[in] response The #az_iot_provisioning_client_register_response obtained after a successful
- *                     call to az_iot_provisioning_client_parse_received_topic_and_payload().
+ * @param[in] response_operation_status An #az_span of the operation status obtained 
+ *                     after a successful call to 
+ *                     az_iot_provisioning_client_parse_received_topic_and_payload().
  * @param[out] out_operation_status The registration operation status.
  * @return An #az_result value indicating the result of the operation.
  * @retval #AZ_ERROR_UNEXPECTED_CHAR If the string contains an unexpected value.
  */
 AZ_NODISCARD az_result az_iot_provisioning_client_parse_operation_status(
-    az_iot_provisioning_client_register_response* response,
+    az_span response_operation_status,
     az_iot_provisioning_client_operation_status* out_operation_status);
 
 /**