+++ sdk/src/azure/iot/az_iot_provisioning_client.c
@@ -379,7 +379,10 @@ AZ_INLINE az_result az_iot_provisioning_client_parse_payload(
       {
         return AZ_ERROR_ITEM_NOT_FOUND;
       }
-      out_response->operation_status = jr.token.slice;
+      _az_RETURN_IF_FAILED(az_iot_provisioning_client_parse_operation_status(
+          jr.token.slice,
+          &out_response->operation_status));
+
       found_operation_status = true;
     }
     else if (az_json_token_is_text_equal(&jr.token, AZ_SPAN_FROM_STR("registrationState")))
@@ -430,7 +433,7 @@ AZ_INLINE az_result az_iot_provisioning_client_parse_payload(
   if (!(found_operation_status && found_operation_id))
   {
     out_response->operation_id = AZ_SPAN_EMPTY;
-    out_response->operation_status = AZ_SPAN_FROM_STR("failed");
+    out_response->operation_status = AZ_IOT_PROVISIONING_STATUS_FAILED;
 
     if (!found_error)
     {
@@ -516,26 +519,29 @@ AZ_NODISCARD az_result az_iot_provisioning_client_parse_received_topic_and_paylo
 }
 
 AZ_NODISCARD az_result az_iot_provisioning_client_parse_operation_status(
-    az_iot_provisioning_client_register_response* response,
+    az_span response_operation_status,
     az_iot_provisioning_client_operation_status* out_operation_status)
 {
-  if (az_span_is_content_equal(response->operation_status, AZ_SPAN_FROM_STR("assigning")))
+  _az_PRECONDITION_VALID_SPAN(response_operation_status, 0, false);
+  _az_PRECONDITION_NOT_NULL(out_operation_status);
+
+  if (az_span_is_content_equal(response_operation_status, AZ_SPAN_FROM_STR("assigning")))
   {
     *out_operation_status = AZ_IOT_PROVISIONING_STATUS_ASSIGNING;
   }
-  else if (az_span_is_content_equal(response->operation_status, AZ_SPAN_FROM_STR("assigned")))
+  else if (az_span_is_content_equal(response_operation_status, AZ_SPAN_FROM_STR("assigned")))
   {
     *out_operation_status = AZ_IOT_PROVISIONING_STATUS_ASSIGNED;
   }
-  else if (az_span_is_content_equal(response->operation_status, AZ_SPAN_FROM_STR("failed")))
+  else if (az_span_is_content_equal(response_operation_status, AZ_SPAN_FROM_STR("failed")))
   {
     *out_operation_status = AZ_IOT_PROVISIONING_STATUS_FAILED;
   }
-  else if (az_span_is_content_equal(response->operation_status, AZ_SPAN_FROM_STR("unassigned")))
+  else if (az_span_is_content_equal(response_operation_status, AZ_SPAN_FROM_STR("unassigned")))
   {
     *out_operation_status = AZ_IOT_PROVISIONING_STATUS_UNASSIGNED;
   }
-  else if (az_span_is_content_equal(response->operation_status, AZ_SPAN_FROM_STR("disabled")))
+  else if (az_span_is_content_equal(response_operation_status, AZ_SPAN_FROM_STR("disabled")))
   {
     *out_operation_status = AZ_IOT_PROVISIONING_STATUS_DISABLED;
   }